Add comprehensive end-to-end tests for complex scenarios, edge cases, and cURL imports
- Implemented complex scenarios and boundary tests in `complex-scenarios.spec.js` to validate handling of large payloads, unusual HTTP methods, and boundary values.
- Created edge cases tests in `edge-cases.spec.js` focusing on special characters, empty parameters, and JSON body handling with special characters.
- Developed importer validation tests in `importer-validation.spec.js` to ensure correct parsing and execution of complex cURL commands, including handling of query strings and headers.
